+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/workshop/AddNewWorkshopViewModel.kt
@@ -159,8 +159,7 @@ class AddNewWorkshopViewModel @Inject constructor(
                 idUser = _formData.value.idUser,
                 description = _formData.value.description,
                 date = _formData.value.date,
-                image = remoteImageUrl,
-                videoTraining = _formData.value.videoTraining
+                image = remoteImageUrl
             )
 
             postAddNewWorkshop(workshopDto).collect { result ->
@@ -201,9 +200,6 @@ class AddNewWorkshopViewModel @Inject constructor(
         if (currentState.description != newState.description) {
             if (!validateInput(newState.description, regexValidation, maxLength = 400)) return
         }
-        if (currentState.videoTraining != newState.videoTraining) {
-            if (!validateInput(newState.name,urlTypingRegex, 100)) return
-        }
         _formData.update { it.update() }
         clearFieldErrors()
     }
@@ -241,13 +237,6 @@ class AddNewWorkshopViewModel @Inject constructor(
         if (data.date.isBlank()) errors["date"] = true
         if (data.description.isBlank()) errors["description"] = true
         if (data.idUser.isBlank()) errors["idUser"] = true
-        if (data.videoTraining.isBlank()) {
-            errors["videoTraining"] = true
-        } else {
-            if (!isValidUrl(data.videoTraining)) {
-                errors["videoTraining"] = true
-            }
-        }
 
         if (data.startHour.isNotBlank() && !hourRegex.matches(data.startHour)) {
             errors["startHour"] = true