+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/workshop/ModifyWorkshopViewModel.kt
@@ -162,8 +162,7 @@ class ModifyWorkshopViewModel @Inject constructor(
                             date = cleanDate ?: "",
                             description = workshopData.description ?: "",
                             idUser = workshopData.idUser ?: "",
-                            image = workshopData.url ?: "",
-                            videoTraining = workshopData.videoTraining ?: ""
+                            image = workshopData.url ?: ""
                         )
                     }
                     _formattedDate.value = formatWorkshopDate(workshopData.date)
@@ -250,8 +249,7 @@ class ModifyWorkshopViewModel @Inject constructor(
                 idUser = _formData.value.idUser,
                 description = _formData.value.description,
                 date = _formData.value.date,
-                image = remoteImageUrl,
-                videoTraining = _formData.value.videoTraining
+                image = remoteImageUrl
             )
             Log.d("WORKSHOP_DEBUG", "Fecha: ${_formData?.value?.date}")
             postModifyWorkshop(workshopDto).collect { result ->
@@ -295,9 +293,6 @@ class ModifyWorkshopViewModel @Inject constructor(
         if (currentState.description != newState.description) {
             if (!validateInput(newState.name,regexValidation, 400)) return
         }
-        if (currentState.videoTraining != newState.videoTraining) {
-            if (!validateInput(newState.name,urlTypingRegex, 100)) return
-        }
         _formData.update { it.update() }
         clearFieldErrors()
     }
@@ -334,14 +329,6 @@ class ModifyWorkshopViewModel @Inject constructor(
         if (data.date.isBlank()) errors["date"] = true
         if (data.description.isBlank()) errors["description"] = true
         if (data.idUser.isBlank()) errors["idUser"] = true
-
-        if (data.videoTraining.isBlank()) {
-            errors["videoTraining"] = true
-        } else {
-            if (!isValidUrl(data.videoTraining)) {
-                errors["videoTraining"] = true
-            }
-        }
 
         if (data.startHour.isNotBlank() && !hourRegex.matches(data.startHour)) {
             errors["startHour"] = true