+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/workshop/WorkshopDetail.kt
@@ -62,45 +62,31 @@ fun WorkshopDetailScreen(
     val errorMessage by viewModel.errorMessage.collectAsState()
     val formattedDate by viewModel.formattedDate.collectAsState()
     ArcaByOlimpoTheme(darkTheme = true, dynamicColor = false) {
-    // State of the snackbar that will handle all messages on the screen.
     val snackbarHostState = remember { SnackbarHostState() }
-    // We use it to launch the coroutine that displays the snackbar.
     val scope = rememberCoroutineScope()
-    // We collect the state exposed by the ViewModel, reacting to changes
     val uiState by viewModel.uiState.collectAsStateWithLifecycle()
 
-
-    // In this case, if there is no error in the uiState,
-    // we consider it a success.
     val haserror = if (uiState.error == null){
         true
     }else {
         false
     }
 
-    // Effect triggered when the uiState.snackbarVisible flag changes.
-    // If true, the snackbar is displayed and the system waits to dismiss it.
     LaunchedEffect(uiState.snackbarVisible) {
 
         if (uiState.snackbarVisible == true){
-            // show snackbar as a suspend function
             scope.launch {
 
                 val result =
                     snackbarHostState.showSnackbar(
                         SnackbarVisualsWithError(
                             "Taller Borrado Correctamente",
-                            // isError controls the visual style (red/blue)
                             isError = true,
                         ),
                     )
-                // When the snack bar ends, we act according to the result
                 when (result){
                     SnackbarResult.Dismissed -> {
-                        //Tell the VM that the visible snackbar is finished
                         viewModel.onSnackbarShown()
-
-                        //Navigate back as soon as the snack bar is finished
                         navController.popBackStack()
 
                     }
@@ -113,15 +99,11 @@ fun WorkshopDetailScreen(
             viewModel.loadWorkshop()
         }
 
-
-    // Displays a confirmation dialog when the user wants to delete an item.
     if (uiState.decisionDialogVisible == true){
         DecisionDialog(
-            // The dialogue closes without any action being taken.
             onDismissRequest = {
                 viewModel.toggledecisionDialog(showdecisionDialog = false)
             },
-            // If the user confirms, the ViewModel is called to delete the input
             onConfirmation = {
                 viewModel.deleteWorkshops(workshopId)
             },
@@ -133,14 +115,11 @@ fun WorkshopDetailScreen(
     }
 
         Scaffold(
-            // Host del snackbar para la pantalla.
         snackbarHost = {
             SnackbarHost(snackbarHostState) { data ->
-                // Every time a snackbar is displayed, our custom component is called
                 Snackbarcustom(
                     data.visuals.message.toString(),
                     modifier = Modifier,
-                    // the type of snack bar (whether it's red or blue).
                     ifSucces = haserror,
                 )
             }
@@ -221,21 +200,6 @@ fun WorkshopDetailScreen(
                                         text= workshop?.description ?: "Cargando descripción...",
                                         style = MaterialTheme.typography.bodyLarge
                                     )
-                                    Spacer(modifier = Modifier.height(24.dp))
-                                    if (workshop?.videoTraining != null) {
-                                        videoView(
-                                            modifier = Modifier
-                                                .fillMaxWidth()
-                                                .height(200.dp),// Dale una altura fija
-                                            videoUrl = workshop!!.videoTraining!!
-                                        )
-                                    }
-                                    else{
-                                        Text(
-                                            text= "No hay video disponible",
-                                            style = MaterialTheme.typography.bodyLarge
-                                        )
-                                    }
                                     Spacer(modifier = Modifier.height(24.dp))
                                     Text(
                                         text= "Sobre el Taller:",