+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/workshop/WorkshopDetailViewModel.kt
@@ -177,8 +177,4 @@ class WorkshopDetailViewModel @Inject constructor(
         }
     }
 
-
-    //--------------
-
-
 }